# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tennis Torunament API",
        "description": "API que permite simular un torneo de tenis y consultar los resultados.",
        "contact": {
            "email": "egarofalo83@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/players/{id}": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Obtiene los detalles de un jugador",
                "operationId": "dbef515d1b5f749e736bca18efe444e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del jugador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de un jugador",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "skill_level": {
                                                "type": "integer"
                                            },
                                            "gender": {
                                                "type": "string"
                                            },
                                            "skills": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string"
                                                        },
                                                        "score": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "tournaments": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "name": {
                                                            "type": "string"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "example": "13-01-2025 16:30:00"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "13-01-2025 16:30:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}